refactor(utils): init cookie extracter & refactor utils
