+++ src/app/category/category.entity.ts
@@ -8,7 +8,7 @@ import {
 import { ObjectType, ID, Field } from '@nestjs/graphql'
 
 import { Product } from '../product/product.entity'
-import { Paginated } from '../../core/utils/params/pagination'
+import { Paginated } from '../../core/utils'
 
 @Entity()
 @ObjectType()
