+++ src/core/utils/lib/index.ts
@@ -1,0 +1,1 @@
+export { extractCookie } from './cookie'
