+++ src/core/utils/params/index.ts
@@ -1,3 +1,3 @@
 export { FilterParams, PaginationFilterArgs } from './filters'
-export { PaginationParams, PaginationArgs } from './pagination'
+export { PaginationParams, PaginationArgs, Paginated } from './pagination'
 export { PaginationSearchParams, PaginationSearchQueryArgs } from './search'
